Code quality cleanup; add MCP demo & storage
Large refactor focused on lint and type cleanup plus new demo/storage features.

- Code quality: removed unused imports/vars, renamed unused locals with _ prefix, removed redundant type assertions, and disabled specific viem lint warnings in key files. Fixed promise handling by adding void where appropriate and converting some async callbacks to sync wrappers to satisfy eslint rules.
- Hooks & logic: minor fixes across many hooks (useActivityFeed, useAgentHealth, useAgents, useJobs, useProposals, etc.) to address floating promises, unnecessary return types, and improved caching/usage. Several components/pages had unused state/props removed or simplified (dashboard, jobs, marketplace, skills, leaderboard, identity, governance, providers, notifications, review, marketplace/create, etc.).
- New features/files: added MCP interactive demo (components/MCPDemoPanel.tsx) with mock data (lib/mcp/mock-data.ts) and simplified local storage for webhooks and push subscriptions (lib/webhooks/storage.ts, lib/push/storage.ts). Also added API types (lib/types/api.ts).
- UI/UX tweaks: small UI handler updates to call functions via void in onClick/onSubmit, removed stray console.log, and simplified some component imports.

This commit is primarily non-functional cleanup to improve type safety, satisfy linters, and add a self-contained MCP demo + persistent webhook/push storage for development and testing.

+++ apps/web/lib/hooks/useChainlinkPrice.ts
@@ -23,10 +23,10 @@ export function useChainlinkEthUsdPrice() {
 
   const priceData: ChainlinkPriceData | null = data
     ? {
-        price: data[1] as bigint, // answer
-        decimals: 8, // ETH/USD typically has 8 decimals
-        updatedAt: Number(data[3]), // updatedAt
-        priceInUsd: Number(data[1]) / 1e8, // Convert to USD
+        price: data[1],
+        decimals: 8,
+        updatedAt: Number(data[3]),
+        priceInUsd: Number(data[1]) / 1e8,
       }
     : null;
